requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
Flask==3.0.3
flatbuffers==24.3.25
humanfriendly==10.0
hydrus-api==5.0.1
idna==3.7
imageio==2.35.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
modules==1.0.0
mpmath==1.3.0
numpy==2.1.0
onnx==1.16.2
onnxruntime==1.19.0
opencv-contrib-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.3
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sympy==1.13.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
